+++ docs/ChangeLog.md
@@ -2,9 +2,11 @@
 
 ## 0.0.4 Alpha (2019-10-22)
 
-* The `splunk/splunk-operator` base image now uses Red Hat's [Universal Base Image](https://developers.redhat.com/products/rhel/ubi/) v8
-* All Enterprise containers are now run using the unprivileged `splunk` user and group
-* Minimum Splunk Enterprise version required is now 8.0
+* Updates to SplunkEnterprise objects are now handled, enabling deployments to be upgraded and modified after creation
+* Added liveness and readiness probes to all Splunk Enterprise pods
+* All Splunk Enterprise containers now run using the unprivileged `splunk` user and group
+* The minimum required version of Splunk Enterprise is now 8.0
+* The `splunk-operator` container now uses Red Hat's [Universal Base Image](https://developers.redhat.com/products/rhel/ubi/) version 8
 
 ## 0.0.3 Alpha (2019-08-14)